Refactor SaleTabs component
